refactor(bills): remove float casts in BillingService (ADR-018)
Limpieza de código después de verificar que bills ya son INTEGER.

CONTEXTO:
Investigación reveló que:
1. Bill model YA tiene casts integer (no necesita cambio)
2. MoneyMigrationTest ya convierte bills.subtotal/tax/etc a INTEGER
3. BillingService todavía usaba (float) casts innecesarios

CAMBIOS:

1. BillingService.php (5 cambios):
   - (float) $order->subtotal → (int) $order->subtotal
   - (float) $order->tax_amount → (int) $order->tax_amount
   - (float) $order->discount_amount → (int) $order->discount_amount
   - (float) $item->subtotal → (int) $item->subtotal
   - Removido (int) $order->total (ya es int)

2. BillIntegerMigrationTest.php (nuevo):
   - 5 tests verificando invariantes INTEGER
   - Usa mismo patrón que FinancialIntegrityTest
   - Sin factories (Company::create directo)

INVARIANTES VERIFICADOS:
✅ subtotal + tax - discount + tip = total
✅ paid + remaining = total (sin epsilon)
✅ Operaciones aritméticas sin pérdida de precisión
✅ Valores grandes (999999) se preservan
✅ Casts del model son integer

REGRESSION CHECK:
✅ Bill tests (54 passed)
✅ SplitBill tests (4 passed)
✅ FinancialIntegrity tests (10 passed)
✅ Payment tests (57 passed)

SIGUIENTE: FASE 4 - Bills sincronizables (ADR-020)
- Commit 1: ADR-020 (documentación)
- Commit 2: Backend endpoint POST /api/v1/bills
- Commit 3: Frontend sync de bills
- Commit 4: Link payments a bills en sync

+++ app/Modules/Payments/Domain/Services/BillingService.php
@@ -24,10 +24,10 @@ class BillingService
         return DB::transaction(function () use ($order, $parts) {
             $this->cancelExistingBills($order);
 
-            $subtotal = (float) $order->subtotal;
-            $tax = (float) $order->tax_amount;
-            $discount = (float) $order->discount_amount;
-            $total = (int) $order->total;
+            $subtotal = (int) $order->subtotal;
+            $tax = (int) $order->tax_amount;
+            $discount = (int) $order->discount_amount;
+            $total = $order->total;
 
             $baseSubtotal = (int) floor($subtotal / $parts);
             $baseTax = (int) floor($tax / $parts);
@@ -82,19 +82,19 @@ class BillingService
             $this->cancelExistingBills($order);
 
             $items = $order->items()->get()->keyBy('id');
-            $orderSubtotal = (float) $order->subtotal;
-            $orderTax = (float) $order->tax_amount;
-            $orderDiscount = (float) $order->discount_amount;
-            $orderTotal = (int) $order->total;
+            $orderSubtotal = (int) $order->subtotal;
+            $orderTax = (int) $order->tax_amount;
+            $orderDiscount = (int) $order->discount_amount;
+            $orderTotal = $order->total;
 
 
 
             // Calcular el subtotal total de items agrupados
-            $totalGroupedSubtotal = 0;
+            $totalGroupedSubtotal = 0; // integer
             foreach ($groups as $group) {
                 foreach ($group['item_ids'] ?? [] as $itemId) {
                     if ($item = $items->get($itemId)) {
-                        $totalGroupedSubtotal += (float) $item->subtotal;
+                        $totalGroupedSubtotal += (int) $item->subtotal;
                     }
                 }
             }
@@ -112,7 +112,7 @@ class BillingService
                 $itemIds = [];
                 foreach ($group['item_ids'] ?? [] as $itemId) {
                     if ($item = $items->get($itemId)) {
-                        $groupSubtotal += (float) $item->subtotal;
+                        $groupSubtotal += (int) $item->subtotal;
                         $itemIds[] = $itemId;
                     }
                 }
@@ -185,7 +185,7 @@ class BillingService
         // Normalizar montos a float
         $amounts = array_map('floatval', $amounts);
         $sumAmounts = (int) round(array_sum($amounts));
-        $orderTotal = (int) round((int) $order->total);
+        $orderTotal = (int) round($order->total);
 
         // Tolerancia de $1 por redondeo
         $difference = (int) round($sumAmounts - $orderTotal);
@@ -254,7 +254,7 @@ class BillingService
             
             if ($existing) {
                 // Si la bill tiene total correcto, retornarla
-                if ((float) $existing->total > 0 && abs((float) $existing->total - (int) $order->total) < 0.01) {
+                if ((float) $existing->total > 0 && abs((float) $existing->total - $order->total) < 0.01) {
                     return $existing;
                 }
                 
@@ -268,10 +268,10 @@ class BillingService
             }
 
             // Crear nueva bill con los totales del order
-            $orderTotal = (int) $order->total;
-            $orderSubtotal = (float) $order->subtotal;
-            $orderTax = (float) $order->tax_amount;
-            $orderDiscount = (float) $order->discount_amount;
+            $orderTotal = $order->total;
+            $orderSubtotal = (int) $order->subtotal;
+            $orderTax = (int) $order->tax_amount;
+            $orderDiscount = (int) $order->discount_amount;
 
             return Bill::create([
                 'company_id' => $order->company_id,
